+++ Nos/Assets/Localization/Localized.swift
@@ -168,6 +168,8 @@ enum Localized: String, Localizable, CaseIterable {
     case invalidKey = "Invalid Key"
     case couldNotReadPrivateKeyMessage = "Could not read your private key. Please verify that it is in nsec or hex format."
     case createAccount = "Create an account"
+    case shareLogs = "Share logs"
+    case failedToExportLogs = "Failed to export logs."
     
     case post = "Post" // (verb form)
     case newNote = "New Note"