+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added a button to share the application logs in the Settings menu
+- Automatically attach debug logs to support emails
+
 ## [0.1 (2)] - 2023-04-14Z
 
 - More performance improvements on the Home tab.